Docs: DTO 문서 위치 수정
