+++ moongge/src/main/java/com/narsha/moongge/controller/UserController.java
@@ -3,9 +3,8 @@ package com.narsha.moongge.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.gson.JsonObject;
 import com.narsha.moongge.base.code.ResponseCode;
-import com.narsha.moongge.base.dto.group.JoinGroupDTO;
+import com.narsha.moongge.group.dto.JoinGroupDTO;
 import com.narsha.moongge.base.dto.response.ResponseDTO;
-import com.narsha.moongge.base.dto.s3.S3FileDTO;
 import com.narsha.moongge.base.dto.user.UpdateUserProfileDTO;
 import com.narsha.moongge.base.dto.user.UserLoginDTO;
 import com.narsha.moongge.base.dto.user.UserRegisterDTO;
