+++ moongge/src/main/java/com/narsha/moongge/group/dto/UpdateTimeDTO.java
@@ -1,17 +1,22 @@
-package com.narsha.moongge.base.dto.group;
+package com.narsha.moongge.group.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.antlr.v4.runtime.misc.NotNull;
 
+import java.time.LocalDateTime;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class JoinGroupDTO {
+public class UpdateTimeDTO {
 
     @NotNull
-    private String userId; // 유저 id
+    private LocalDateTime startTime; // 시작 시간
+
+    @NotNull
+    private LocalDateTime endTime; // 끝 시간
 
     @NotNull
     private String groupCode; // 그룹 코드
